Optionally request MANAGE_EXTERNAL_STORAGE for write access
Due to the scoped storage constraints, the only way to get a publicly
writable directory with a real, non-content-resolver-based, path is
via the MANAGE_EXTERNAL_STORAGE permission.

The mitm addon now allow the user to optionally request it. When
granted, user addons will be loaded directly from the public dir
without an additional copy.

Closes #16